refactor(checks): extract prop-driven ChecksPanelInner from active-worktree wrapper
Split ChecksPanel into a thin active-worktree wrapper (default export) and a
prop-driven ChecksPanelInner that reads no active-worktree globals. The wrapper
supplies worktree/repo/gitIdentityDisplay/linked-review identity as explicit
props; the inner is agnostic about whether its target is the active worktree, so
a shipped-PR caller (Mission Control, PR2) can drive the same UI from a branch/PR
identity without a live worktree.

Also extracted leaf modules: ChecksPanelReviewHeader, GitLab review RPC helpers,
and the inner prop/types. Behavior-preserving for the existing Checks tab.

(In progress) max-lines split of ChecksPanelInner still pending; temporary
eslint-disable retained until the hook/component extraction lands.
