+++ src/renderer/src/components/right-sidebar/ChecksPanelReviewHeader.tsx
@@ -1,0 +1,112 @@
+import React from 'react'
+import { RefreshCw, Ellipsis, Link, Unlink, GitMerge } from 'lucide-react'
+import { cn } from '@/lib/utils'
+import { Button } from '@/components/ui/button'
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger
+} from '@/components/ui/dropdown-menu'
+import { getTerminalUrlSystemBrowserHint } from '../terminal-pane/terminal-link-open-hints'
+import { PullRequestIcon, prStateColor } from './checks-panel-content'
+import { type ChecksPanelReview } from './checks-panel-review'
+import { translate } from '@/i18n/i18n'
+
+type ChecksPanelReviewHeaderProps = {
+  review: ChecksPanelReview
+  isRefreshing: boolean
+  canUnlinkPullRequest: boolean
+  showSystemBrowserHint: boolean
+  onRefresh: () => void
+  onOpenReview: (event: React.MouseEvent<HTMLButtonElement>) => void
+  onUnlinkPullRequest: () => void
+  onLinkAnotherPullRequest: () => void
+}
+
+export function ChecksPanelReviewHeader({
+  review,
+  isRefreshing,
+  canUnlinkPullRequest,
+  showSystemBrowserHint,
+  onRefresh,
+  onOpenReview,
+  onUnlinkPullRequest,
+  onLinkAnotherPullRequest
+}: ChecksPanelReviewHeaderProps): React.JSX.Element {
+  const reviewNumberLabel = review.provider === 'gitlab' ? `!${review.number}` : `#${review.number}`
+  const ReviewIcon = review.provider === 'gitlab' ? GitMerge : PullRequestIcon
+  const reviewHostLabel = review.provider === 'gitlab' ? 'GitLab' : 'GitHub'
+  const showPullRequestMenu = review.provider === 'github'
+  const openTitle = translate(
+    'auto.components.right.sidebar.ChecksPanel.5c88c6db07',
+    'Open on {{value0}}',
+    { value0: reviewHostLabel }
+  )
+  const title = showSystemBrowserHint
+    ? `${openTitle}. ${getTerminalUrlSystemBrowserHint()}`
+    : openTitle
+
+  return (
+    <div className="flex items-center gap-2">
+      <ReviewIcon className="size-4 text-muted-foreground shrink-0" />
+      <button
+        type="button"
+        className="rounded px-0.5 text-[12px] font-semibold text-foreground underline decoration-border underline-offset-2 hover:text-foreground hover:decoration-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring"
+        title={title}
+        onClick={onOpenReview}
+      >
+        {reviewNumberLabel}
+      </button>
+      <span
+        className={cn(
+          'text-[9px] font-semibold uppercase tracking-wider px-1.5 py-0.5 rounded border',
+          prStateColor(review.state)
+        )}
+      >
+        {review.state}
+      </span>
+      <div className="flex-1" />
+      <button
+        className="cursor-pointer rounded p-1 text-muted-foreground transition-colors hover:bg-accent hover:text-foreground disabled:cursor-default disabled:opacity-50"
+        title={translate('auto.components.right.sidebar.ChecksPanel.7f4489f370', 'Refresh')}
+        onClick={onRefresh}
+        disabled={isRefreshing}
+      >
+        <RefreshCw className={cn('size-3.5', isRefreshing && 'animate-spin')} />
+      </button>
+      {showPullRequestMenu && (
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button
+              type="button"
+              variant="ghost"
+              size="icon-xs"
+              aria-label={translate(
+                'auto.components.right.sidebar.ChecksPanel.653c105ecc',
+                'More PR actions'
+              )}
+              title={translate(
+                'auto.components.right.sidebar.ChecksPanel.653c105ecc',
+                'More PR actions'
+              )}
+              className="text-muted-foreground hover:text-foreground"
+            >
+              <Ellipsis className="size-3.5" />
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent align="end" className="w-44">
+            <DropdownMenuItem disabled={!canUnlinkPullRequest} onSelect={onUnlinkPullRequest}>
+              <Unlink className="size-3.5" />
+              {translate('auto.components.right.sidebar.ChecksPanel.7202f4a40a', 'unlink PR')}
+            </DropdownMenuItem>
+            <DropdownMenuItem onSelect={onLinkAnotherPullRequest}>
+              <Link className="size-3.5" />
+              {translate('auto.components.right.sidebar.ChecksPanel.07871c0589', 'Link another PR')}
+            </DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+      )}
+    </div>
+  )
+}
