+++ src/renderer/src/components/right-sidebar/checks-panel-gitlab-review-rpc.ts
@@ -1,0 +1,82 @@
+import type {
+  GitLabDiscussionResolveResult,
+  GitLabWorkItemDetails,
+  PRComment
+} from '../../../../shared/types'
+import { callRuntimeRpc, getActiveRuntimeTarget } from '@/runtime/runtime-rpc-client'
+import type { ChecksPanelReview } from './checks-panel-review'
+
+export function isGitLabChecksPanelReview(
+  review: ChecksPanelReview | null
+): review is ChecksPanelReview & { provider: 'gitlab' } {
+  return review?.provider === 'gitlab'
+}
+
+export function gitLabMRCommentsToPRComments(
+  comments: GitLabWorkItemDetails['comments'] | undefined
+): PRComment[] {
+  return (comments ?? []).map((comment) => {
+    const { reactions: _reactions, ...compatibleComment } = comment
+    // Why: the shared comments renderer expects GitHub reaction content enums;
+    // GitLab emoji award names are open-ended, so omit them in this view.
+    return compatibleComment
+  })
+}
+
+export async function fetchGitLabMRDetailsForChecks(args: {
+  repoPath: string
+  repoId?: string
+  settings: Parameters<typeof getActiveRuntimeTarget>[0]
+  iid: number
+}): Promise<GitLabWorkItemDetails | null> {
+  const target = getActiveRuntimeTarget(args.settings)
+  if (target.kind === 'environment') {
+    return callRuntimeRpc<GitLabWorkItemDetails | null>(
+      target,
+      'gitlab.workItemDetails',
+      {
+        repo: args.repoId ?? args.repoPath,
+        iid: args.iid,
+        type: 'mr'
+      },
+      { timeoutMs: 30_000 }
+    )
+  }
+  return (await window.api.gl.workItemDetails({
+    repoPath: args.repoPath,
+    repoId: args.repoId,
+    iid: args.iid,
+    type: 'mr'
+  })) as GitLabWorkItemDetails | null
+}
+
+export async function resolveGitLabMRDiscussionForChecks(args: {
+  repoPath: string
+  repoId?: string
+  settings: Parameters<typeof getActiveRuntimeTarget>[0]
+  iid: number
+  discussionId: string
+  resolved: boolean
+}): Promise<GitLabDiscussionResolveResult> {
+  const target = getActiveRuntimeTarget(args.settings)
+  if (target.kind === 'environment') {
+    return callRuntimeRpc<GitLabDiscussionResolveResult>(
+      target,
+      'gitlab.resolveMRDiscussion',
+      {
+        repo: args.repoId ?? args.repoPath,
+        iid: args.iid,
+        discussionId: args.discussionId,
+        resolved: args.resolved
+      },
+      { timeoutMs: 30_000 }
+    )
+  }
+  return window.api.gl.resolveMRDiscussion({
+    repoPath: args.repoPath,
+    repoId: args.repoId,
+    iid: args.iid,
+    discussionId: args.discussionId,
+    resolved: args.resolved
+  })
+}
